tools/test_apps/system/test_watchpoint/pytest_watchpoint.py
# SPDX-FileCopyrightText: 2023 Espressif Systems (Shanghai) CO LTD
# SPDX-License-Identifier: CC0-1.0

import pytest
from pytest_embedded import Dut


@pytest.mark.esp32
@pytest.mark.esp32c3
@pytest.mark.generic
def test_watchpoint(dut: Dut) -> None:
    dut.expect_exact('stacks clean', timeout=3)
